func_detach
